Implement USER/handshake. Beta Privmsg reveals deadlock issues.
Work on switching to RwLock instead of Mutex. Also noticed
one probable cause of the current deadlock bug: the polling fn
in ClientFuture was holding an RwLock on the Client Structure
during the whole time the thread was working through parser
and command() etc. Trying to move to more atomic use of locks,
e.g. not just when a function is strictly speaking mut borrow,
but for the smallest segments of code where actual writing
occurs. In the case of the client handling code, much of this
is in src/buffer.rs and it is not necessary for a given client
to be locked for the entire duration of poll() and every other
bit of the program, since that's essentially the entry point.

Signed-off-by: aoei/Jo