Fix nick-hoarding bug
It was noticed that after a second NICK command, both old and
new nicks were effective as targets for another user to send
PRIVMSGs. Fix by removing the old hash key when a nick is changed.

Change error handling code in order to send text-based error replies
to the client. Clients are no longer killed due to erroneous commands,
but no error message is being sent back. Not sure why.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>