Return a Vec of reply/error messages to client
Define clearer separation between error messages for the client and
error messages for the internal running of the ircd.
Add additional IRC welcome replies 002 - 004 with hostname info, version
info and other details about the server.
IRC command handling functions now return a Vec of messages for the
client, which will propagate up the call-stack and be handled in
client.rs. Only messages sent to other clients need to be handled with
side-effects in that sense.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>